Allow init flags to bypass questions
