+++ pkg/cli/initialize/initialize_fn_test.go
@@ -1,4 +1,4 @@
-package cli
+package initialize
 
 import (
 	"context"
